Handle creation of jobs in AddBuilding instead of nobUsual ctor
For the same reason why AddBuilding is not called in the noBuilding
ctor. The inspiration for this fix came from #1048, which is an
assertion triggered by the fact that nobUsual::nobUsual creates the
worker for the building, which in case of a shipyard is a
nofShipWright; the nofShipWright constructor asserts that the
building is, in fact, a nobShipYard. It does that by using a
dynamic_cast, which is undefined behaviour, since the constructor
for nobShipYard has not started running yet (remember, the worker
is created in the inherited nobUsual ctor).

See also https://stackoverflow.com/a/6299463/1248008

Fixes #1048.