Fix crash when destroying harbor with an ordered builder for an expedition
When ordering a builder it is assumed that it is put into the jobsWanted
list which does not happen if the builder can be sent right away. On
destruction the jobWanted is not found and the end-iterator is passed to
list::erase which leads to corruption of the list.

Fixes #1071