Handle index mapping for 0-range
If we’re mapping a value from a zero range, e.a. min(range) = max(range) then automatically choose zero index. This might be an arbitrary choice but works for all use cases of the function without an error. Could be changed to use another specific value, e.g. a random value (which requires a random map generator) or just the middle value (which requires additional computation).