Split GameWorldView::MoveTo into MoveTo and MoveBy
The default is to move relative so "move to" doesn't make sense.
As we don't need dynamic change of this behavior we can use 2 functions
for that with appropritate names.
Also remove the MoveToX/MoveToY functions as the same can be done by the
other 2 functions and they suffer from the same problem.