Fix potential crash in glFont ctor
glFont relies on the textures being a `glArchivItem_Bitmap` but uses the
libsiedler2 allocator which might not be set to `glAllocator`.
This happens e.g. in `testControls` depending on the test order as some
tests do install the `glAllocator` so the other tests using e.g.
`createMockFont` succeed only if run after the former.

Just create the required bitmap type explicitly