Handle situation when a farmhand arrives at a target that became unvailable
We keep checking for validity as long as we are walking. But when we
arrive we didn't check again.
So when the point became invalid (e.g. out of resources) during the time
it took from walking from starting to walk from the neighbour node until
arrival or if the point becomes invalid in the same GF of arrival we
started working at that invalid point.

Move the validity check to the top of the function.
Also use `WorkAborted` to reset the reserved state as a derived class
might do more in this case.