Fix missing message box when map fails to load
`StartGame` calls `CI_GameLoading` which creates the load screen
desktop.
Directly afterwards the map or savegame is loaded which may fail.
This calls the error callback causing the load screen to show a message
box which on confirming switches back to the main menu.
However at this point the desktop is not yet active so the message box
will be show for the previous desktop and closed right when the current
one is shown.
Hence check if we are active yet and call ShowAfterSwitch otherwise.