Fix compatibility for Boost 1.81+
`boost::filesystem::path` allows conversion from anything that is
convertible to a string-ish type, like C-Strings or `std::string`.
`kaguya::LuaRef` is convertible to everything (but may fail at runtime)
which makes the conversion to that ambiguous.
This becomes a problem for `ImageResource` which has an implicit
conversion from `boost::filesystem::path` so implicitely converting
`kaguya::LuaRef` to `ImageResource` will make the compiler try the
conversion to `boost::filesystem::path` causing an error.

Use the `get<T>` member function to avoid the atempted implicit conversion.