Add functions to search for terrain descriptors
In some places we need e.g. "any water" terrain or in general a single
or all descriptors matching a condition.
Add `find` and `findAll` methods to the containers and add and use the
explicit bool operator.
Both in combination avoid using the raw `value` member of the `DescIdx`
in most cases.