Merge `dskCampaignMissionMapSelection` and `dskCampaignMissionSelection`
The classes share some code and essentially the same functionality.
Especially problematic is the reliance on the "mission map" in the
campagin description which may lead to an empty screen if
`dskCampaignMissionMapSelection` is instantiated in absence of that.

Additionally the difference in naming doesn't make it clear which is which.